Check Method Signatures in External Refinements (#76)

+++ liquidjava-example/src/main/java/testSuite/math/correctInvocation/MathRefinements.java
@@ -16,13 +16,13 @@ public interface MathRefinements {
     public int abs(int arg0);
 
     @Refinement("(arg0 > 0)?( _ == arg0):(_ == -arg0)")
-    public int abs(long arg0);
+    public long abs(long arg0);
 
     @Refinement("(arg0 > 0)?( _ == arg0):(_ == -arg0)")
-    public int abs(float arg0);
+    public float abs(float arg0);
 
     @Refinement("(arg0 > 0)?( _ == arg0):(_ == -arg0)")
-    public int abs(double arg0);
+    public double abs(double arg0);
 
     @Refinement(" _ == a+b")
     public int addExact(int a, int b);
@@ -43,13 +43,13 @@ public interface MathRefinements {
     public int decrementExact(int a);
 
     @Refinement("_ == (a-1)")
-    public int decrementExact(long a);
+    public long decrementExact(long a);
 
     @Refinement("_ == (a+1)")
     public int incrementExact(int a);
 
     @Refinement("_ == (a+1)")
-    public int incrementExact(long a);
+    public long incrementExact(long a);
 
     @Refinement("(a > b)? (_ == a):(_ == b)")
     public int max(int a, int b);
